// src/common/schema.ts
export type Maybe<T> = T | undefined;
// ...
export type CurrencyCode = typeof CurrencyCode[keyof typeof CurrencyCode];
// ...
export type HasMetafields = {
	/** Returns a metafield found by namespace and key. */
	metafield?: Maybe<Metafield>;
	/**
	 * A paginated list of metafields associated with the resource.
	 * @deprecated As of 2022-07, the paginated `metafields` field has been repurposed to require a list of metafield identifiers.
	 *
	 */
	metafields: MetafieldConnection;
};
// ...
export type ImageConnection = {
	__typename?: 'ImageConnection';
	/** A list of edges. */
	edges: ImageEdge[];
	/** A list of the nodes contained in ImageEdge. */
	nodes: Image[];
	/** Information to aid in pagination. */
	pageInfo: PageInfo;
};

/** List of supported image content types. */
export const ImageContentType = {
	/** A JPG image. */
	Jpg: 'JPG',
	/** A PNG image. */
	Png: 'PNG',
	/** A WEBP image. */
	Webp: 'WEBP',
} as const;

export type ImageContentType =
	typeof ImageContentType[keyof typeof ImageContentType];
/**
 * An auto-generated type which holds one Image and a cursor during pagination.
 *
 */
export type ImageEdge = {
	__typename?: 'ImageEdge';
	/** A cursor for use in pagination. */
	cursor: Scalars['String'];
	/** The item at the end of ImageEdge. */
	node: Image;
};
// ...
export type MoneyV2 = {
	__typename?: 'MoneyV2';
	/** Decimal money amount. */
	amount: Scalars['Decimal'];
	/** Currency of the money. */
	currencyCode: CurrencyCode;
};
// ...
export type Node = {
	/** A globally-unique identifier. */
	id: Scalars['ID'];
};
// ...
export type PageInfo = {
	__typename?: 'PageInfo';
	/** The cursor corresponding to the last node in edges. */
	endCursor?: Maybe<Scalars['String']>;
	/** Whether there are more pages to fetch following the current page. */
	hasNextPage: Scalars['Boolean'];
	/** Whether there are any pages prior to the current page. */
	hasPreviousPage: Scalars['Boolean'];
	/** The cursor corresponding to the first node in edges. */
	startCursor?: Maybe<Scalars['String']>;
};
// ...
export type Product = HasMetafields &
	Node &
	OnlineStorePublishable & {
		__typename?: 'Product';
		/** Indicates if at least one product variant is available for sale. */
		availableForSale: Scalars['Boolean'];
		/** List of collections a product belongs to. */
		collections: CollectionConnection;
		/** The compare at price of the product across all variants. */
		compareAtPriceRange: ProductPriceRange;
		/** The date and time when the product was created. */
		createdAt: Scalars['DateTime'];
		/** Stripped description of the product, single line with HTML tags removed. */
		description: Scalars['String'];
		/** The description of the product, complete with HTML formatting. */
		descriptionHtml: Scalars['HTML'];
		/**
		 * The featured image for the product.
		 *
		 * This field is functionally equivalent to `images(first: 1)`.
		 *
		 */
		featuredImage?: Maybe<Image>;
		/**
		 * A human-friendly unique string for the Product automatically generated from its title.
		 * They are used by the Liquid templating language to refer to objects.
		 *
		 */
		handle: Scalars['String'];
		/** A globally-unique identifier. */
		id: Scalars['ID'];
		/** List of images associated with the product. */
		images: ImageConnection;
		/** The media associated with the product. */
		media: MediaConnection;
		/** Returns a metafield found by namespace and key. */
		metafield?: Maybe<Metafield>;
		/**
		 * A paginated list of metafields associated with the resource.
		 * @deprecated As of 2022-07, the paginated `metafields` field has been repurposed to require a list of metafield identifiers.
		 *
		 */
		metafields: MetafieldConnection;
		/** The URL used for viewing the resource on the shop's Online Store. Returns `null` if the resource is currently not published to the Online Store sales channel. */
		onlineStoreUrl?: Maybe<Scalars['URL']>;
		/** List of product options. */
		options: ProductOption[];
		/** The price range. */
		priceRange: ProductPriceRange;
		/** A categorization that a product can be tagged with, commonly used for filtering and searching. */
		productType: Scalars['String'];
		/** The date and time when the product was published to the channel. */
		publishedAt: Scalars['DateTime'];
		/** Whether the product can only be purchased with a selling plan. */
		requiresSellingPlan: Scalars['Boolean'];
		/** A list of a product's available selling plan groups. A selling plan group represents a selling method. For example, 'Subscribe and save' is a selling method where customers pay for goods or services per delivery. A selling plan group contains individual selling plans. */
		sellingPlanGroups: SellingPlanGroupConnection;
		/** The product's SEO information. */
		seo: Seo;
		/**
		 * A comma separated list of tags that have been added to the product.
		 * Additional access scope required for private apps: unauthenticated_read_product_tags.
		 *
		 */
		tags: Array<Scalars['String']>;
		/** The product’s title. */
		title: Scalars['String'];
		/** The total quantity of inventory in stock for this Product. */
		totalInventory?: Maybe<Scalars['Int']>;
		/**
		 * The date and time when the product was last modified.
		 * A product's `updatedAt` value can change for different reasons. For example, if an order
		 * is placed for a product that has inventory tracking set up, then the inventory adjustment
		 * is counted as an update.
		 *
		 */
		updatedAt: Scalars['DateTime'];
		/**
		 * Find a product’s variant based on its selected options.
		 * This is useful for converting a user’s selection of product options into a single matching variant.
		 * If there is not a variant for the selected options, `null` will be returned.
		 *
		 */
		variantBySelectedOptions?: Maybe<ProductVariant>;
		/** List of the product’s variants. */
		variants: ProductVariantConnection;
		/** The product’s vendor name. */
		vendor: Scalars['String'];
	};
// ...
export type ProductOption = Node & {
	__typename?: 'ProductOption';
	/** A globally-unique identifier. */
	id: Scalars['ID'];
	/** The product option’s name. */
	name: Scalars['String'];
	/** The corresponding value to the product option name. */
	values: Array<Scalars['String']>;
};

/** The price range of the product. */
export type ProductPriceRange = {
	__typename?: 'ProductPriceRange';
	/** The highest variant's price. */
	maxVariantPrice: MoneyV2;
	/** The lowest variant's price. */
	minVariantPrice: MoneyV2;
};
// ...
export type ProductVariant = HasMetafields &
	Node & {
		__typename?: 'ProductVariant';
		/** Indicates if the product variant is available for sale. */
		availableForSale: Scalars['Boolean'];
		/** The barcode (for example, ISBN, UPC, or GTIN) associated with the variant. */
		barcode?: Maybe<Scalars['String']>;
		/**
		 * The compare at price of the variant. This can be used to mark a variant as on sale, when `compareAtPrice` is higher than `price`.
		 * @deprecated Use `compareAtPriceV2` instead
		 */
		compareAtPrice?: Maybe<Scalars['Money']>;
		/** The compare at price of the variant. This can be used to mark a variant as on sale, when `compareAtPriceV2` is higher than `priceV2`. */
		compareAtPriceV2?: Maybe<MoneyV2>;
		/** Whether a product is out of stock but still available for purchase (used for backorders). */
		currentlyNotInStock: Scalars['Boolean'];
		/** A globally-unique identifier. */
		id: Scalars['ID'];
		/**
		 * Image associated with the product variant. This field falls back to the product image if no image is available.
		 *
		 */
		image?: Maybe<Image>;
		/** Returns a metafield found by namespace and key. */
		metafield?: Maybe<Metafield>;
		/**
		 * A paginated list of metafields associated with the resource.
		 * @deprecated As of 2022-07, the paginated `metafields` field has been repurposed to require a list of metafield identifiers.
		 *
		 */
		metafields: MetafieldConnection;
		/**
		 * The product variant’s price.
		 * @deprecated Use `priceV2` instead
		 */
		price: Scalars['Money'];
		/** The product variant’s price. */
		priceV2: MoneyV2;
		/** The product object that the product variant belongs to. */
		product: Product;
		/** The total sellable quantity of the variant for online sales channels. */
		quantityAvailable?: Maybe<Scalars['Int']>;
		/** Whether a customer needs to provide a shipping address when placing an order for the product variant. */
		requiresShipping: Scalars['Boolean'];
		/** List of product options applied to the variant. */
		selectedOptions: SelectedOption[];
		/** Represents an association between a variant and a selling plan. Selling plan allocations describe which selling plans are available for each variant, and what their impact is on pricing. */
		sellingPlanAllocations: SellingPlanAllocationConnection;
		/** The SKU (stock keeping unit) associated with the variant. */
		sku?: Maybe<Scalars['String']>;
		/** The in-store pickup availability of this variant by location. */
		storeAvailability: StoreAvailabilityConnection;
		/** The product variant’s title. */
		title: Scalars['String'];
		/** The unit price value for the variant based on the variant's measurement. */
		unitPrice?: Maybe<MoneyV2>;
		/** The unit price measurement for the variant. */
		unitPriceMeasurement?: Maybe<UnitPriceMeasurement>;
		/** The weight of the product variant in the unit system specified with `weight_unit`. */
		weight?: Maybe<Scalars['Float']>;
		/** Unit of measurement for weight. */
		weightUnit: WeightUnit;
	};
// ...
export type ProductVariantConnection = {
	__typename?: 'ProductVariantConnection';
	/** A list of edges. */
	edges: ProductVariantEdge[];
	/** A list of the nodes contained in ProductVariantEdge. */
	nodes: ProductVariant[];
	/** Information to aid in pagination. */
	pageInfo: PageInfo;
};

/**
 * An auto-generated type which holds one ProductVariant and a cursor during pagination.
 *
 */
export type ProductVariantEdge = {
	__typename?: 'ProductVariantEdge';
	/** A cursor for use in pagination. */
	cursor: Scalars['String'];
	/** The item at the end of ProductVariantEdge. */
	node: ProductVariant;
};
// ...
export type SelectedOption = {
	__typename?: 'SelectedOption';
	/** The product option’s name. */
	name: Scalars['String'];
	/** The product option’s value. */
	value: Scalars['String'];
};
// ...
export type BasicProductFragment = {
	__typename?: 'Product';
	id: string;
	title: string;
	handle: string;
	productType: string;
	vendor: string;
};
// ...
export type FullProductFragment = {
	__typename?: 'Product';
	id: string;
	handle: string;
	availableForSale: boolean;
	title: string;
	productType: string;
	vendor: string;
	description: string;
	descriptionHtml: string;
	totalInventory?: number | undefined;
	options: Array<{
		__typename?: 'ProductOption';
		id: string;
		name: string;
		values: string[];
	}>;
	priceRange: {
		__typename?: 'ProductPriceRange';
		maxVariantPrice: {
			__typename?: 'MoneyV2';
			amount: string;
			currencyCode: CurrencyCode;
		};
		minVariantPrice: {
			__typename?: 'MoneyV2';
			amount: string;
			currencyCode: CurrencyCode;
		};
	};
	variants: {
		__typename?: 'ProductVariantConnection';
		pageInfo: {
			__typename?: 'PageInfo';
			hasNextPage: boolean;
			hasPreviousPage: boolean;
		};
		edges: Array<{
			__typename?: 'ProductVariantEdge';
			cursor: string;
			node: {
				__typename?: 'ProductVariant';
				id: string;
				title: string;
				sku?: string | undefined;
				availableForSale: boolean;
				requiresShipping: boolean;
				quantityAvailable?: number | undefined;
				currentlyNotInStock: boolean;
				selectedOptions: Array<{
					__typename?: 'SelectedOption';
					name: string;
					value: string;
				}>;
				priceV2: {
					__typename?: 'MoneyV2';
					amount: string;
					currencyCode: CurrencyCode;
				};
				compareAtPriceV2?:
					| {
							__typename?: 'MoneyV2';
							amount: string;
							currencyCode: CurrencyCode;
					  }
					| undefined;
				image?:
					| {
							__typename?: 'Image';
							id?: string | undefined;
							width?: number | undefined;
							height?: number | undefined;
							altText?: string | undefined;
							originalSrc: string;
					  }
					| undefined;
			};
		}>;
	};
	images: {
		__typename?: 'ImageConnection';
		pageInfo: {
			__typename?: 'PageInfo';
			hasNextPage: boolean;
			hasPreviousPage: boolean;
		};
		edges: Array<{
			__typename?: 'ImageEdge';
			cursor: string;
			node: {
				__typename?: 'Image';
				id?: string | undefined;
				width?: number | undefined;
				height?: number | undefined;
				altText?: string | undefined;
				originalSrc: string;
			};
		}>;
	};
};